DolphinQt: Reduce maximum initial dialog size to 75% of the screen dimensions.
The maximum size for the initial size in dialogs has been reduced from
90% to 75% of the screen dimensions.

Rationale: Almost no dialog in the application requires such a tall
window resolution (an exception is the **Graphics > Advanced** pane in
the **Settings** dialog). The reduced size prevents dialogs from
appearing close to the edges of the screen, which was _opinionatedly_
not too elegant.

| Before | After |
| ------ | ----- |
| <img width="1920" height="1080" alt="[Dolphin Emulator] Overly long Settings dialog" title="[Dolphin Emulator] Overly long Settings dialog" src="https://github.com/user-attachments/assets/0c5ecded-68bf-46e2-8696-399b2f263345" /> | <img width="1920" height="1080" alt="[Dolphin Emulator] Reduced Settings dialog" title="[Dolphin Emulator] Reduced Settings dialog" src="https://github.com/user-attachments/assets/45a63f2f-a7a7-4663-bacf-96d4172d42bf" /> |